# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marine Corps logistics calculator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /units/{unitId}/weapons:
    get:
      operationId: listUnitWeapons
      tags: [units]
      summary: List weapon systems on a unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of weapon entries for this unit
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UnitWeapon"
    post:
      operationId: addUnitWeapon
      tags: [units]
      summary: Add a weapon system to a unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitWeaponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitWeapon"
  /units/{unitId}/weapons/{weaponEntryId}:
    patch:
      operationId: updateUnitWeapon
      tags: [units]
      summary: Update weapon quantity for a unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
        - name: weaponEntryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitWeaponUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitWeapon"
    delete:
      operationId: deleteUnitWeapon
      tags: [units]
      summary: Remove a weapon system from a unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
        - name: weaponEntryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /weapon-systems/{weaponSystemId}/dodic-rates/{rateId}:
    patch:
      operationId: updateWeaponDodicRate
      tags: [units]
      summary: Edit doctrinal combat-load / assault / sustain rates for a weapon-DODIC pair
      description: Allows planners to adjust doctrinal burn rates per weapon system without touching the database. Changes are immediately reflected in unit-level DOS calculations. The full rate matrix (weapon × DODIC × GCE/Non-GCE) is seeded once by `scripts/src/seed.ts`; only edits are exposed via the API. There is intentionally no list/create/delete endpoint for rate rows.
      parameters:
        - name: weaponSystemId
          in: path
          required: true
          schema:
            type: string
        - name: rateId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeaponDodicRateInput"
      responses:
        "200":
          description: Updated rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeaponDodicRate"
        "400":
          description: Invalid payload
        "404":
          description: Weapon system or rate not found
...
components:
  schemas:
...
    WeaponDodicRate:
      type: object
      properties:
        id:
          type: string
          description: Stable identifier for this weapon-DODIC rate row
        dodic:
          type: string
        nomenclature:
          type: string
        catalogItemId:
          type: string
        gceCombatLoad:
          type: number
          minimum: 0
        gceAssaultRate:
          type: number
          minimum: 0
        gceSustainRate:
          type: number
          minimum: 0
        nonGceCombatLoad:
          type: number
          minimum: 0
        nonGceAssaultRate:
          type: number
          minimum: 0
        nonGceSustainRate:
          type: number
          minimum: 0
      required:
        - id
        - dodic
        - nomenclature
        - catalogItemId
        - gceCombatLoad
        - gceAssaultRate
        - gceSustainRate
        - nonGceCombatLoad
        - nonGceAssaultRate
        - nonGceSustainRate
    WeaponDodicRateInput:
      type: object
      description: Editable doctrinal rates for a weapon-DODIC pair. All fields optional; only provided fields are updated.
      properties:
        gceCombatLoad:
          type: number
          minimum: 0
        gceAssaultRate:
          type: number
          minimum: 0
        gceSustainRate:
          type: number
          minimum: 0
        nonGceCombatLoad:
          type: number
          minimum: 0
        nonGceAssaultRate:
          type: number
          minimum: 0
        nonGceSustainRate:
          type: number
          minimum: 0
    UnitWeapon:
      type: object
      properties:
        id:
          type: string
        unitId:
          type: string
        weaponSystemId:
          type: string
        weaponName:
          type: string
        tamcn:
          type: string
          nullable: true
        quantity:
          type: integer
          minimum: 1
        isGce:
          type: boolean
      required: [id, unitId, weaponSystemId, weaponName, quantity, isGce]
    UnitWeaponInput:
      type: object
      properties:
        weaponSystemId:
          type: string
        quantity:
          type: integer
          minimum: 1
      required: [weaponSystemId, quantity]
    UnitWeaponUpdate:
      type: object
      properties:
        quantity:
          type: integer
          minimum: 1
      required: [quantity]
...